Client moved out of the Clients namespace
